Remove Python 2 support completely
I think this was a bad idea. Even if this program is nothing big at all,
but Python 2/3 compatibility is a PITA even for basic tasks (like a
import).

If you need Python 2 support, you can try 3to2. Not gonna guarantee that
it works though, because I want to use the modern features and libraries
from Python 3.