Better filename generation
Now I get the page HTML, take the page title (between '<title></title>'),
parse it (so it's safe to use as filename) and use it as a filename.

The code is still very raw, need more tests (specially in Windows, since
the filename restrictions in Windows are greater than on Unix), but
should work.

Fix issue #5