Drop support for Python 3.2
This is not because any change to this project, but because upstream
(livestreamer) doesn't support Python 3.2 anymore since version 1.9.0.
Since I always use the lastest version of livestreamer doesn't make
much sense to try to support an older version. And btw, we still support
the much older Python 2.7 so for those that can't update may still use
Python 2.7.